moved UART to specific files & migrated to CMake (for Boost)
