+++ .gitignore
@@ -16,3 +16,14 @@
 
 # Some lib directory
 TM4C/lib
+
+# CMake files
+CMakeCache.txt
+CMakeFiles
+CMakeScripts
+Testing
+Makefile
+cmake_install.cmake
+install_manifest.txt
+compile_commands.json
+CTestTestfile.cmake
