+++ Pi/src/cpp/comm/comm.cpp
@@ -1,14 +1,12 @@
 #include "comm/comm.h"
+#include "comm/uart.h"
 #include "gpio/GPIOPin.h"
+
+#include "boost/asio.hpp"
 
 #include <vector>
 #include <string>
 #include <algorithm>
-#include <errno.h>
-#include <fcntl.h>
-#include <string.h>
-#include <termios.h>
-#include <unistd.h>
 
 using namespace std;
 
@@ -25,64 +23,7 @@ GPIOPin e_pin;//for ending reading command
 vector<GPIOPin> dwrite_block;//for writing to digital outputs
 
 //com api
-int curr_fd;
-
-int set_iface_attribs(int fd, int speed, int parity) {
-	struct termios tty;
-	memset(&tty, 0, sizeof(tty));
-	if (tcgetattr(fd, &tty) != 0) {
-		printf("error %d from tcgetattr\n", errno);
-		return -1;
-	}
-
-	cfsetospeed(&tty, speed);
-	cfsetispeed(&tty, speed);
-
-	tty.c_cflag = (tty.c_cflag & ~CSIZE) | CS8;//8 bit characters
-	tty.c_iflag &= ~IGNBRK;//disable ignore_break
-	tty.c_lflag = 0;//no signaling chars, no echo
-	tty.c_oflag = 0;//no remapping, no delays
-	tty.c_cc[VMIN] = 0;//read doesn't block
-	tty.c_cc[VTIME] = 5;//0.5 second read timeout
-	tty.c_iflag &= ~(IXON | IXOFF | IXANY);//shut off xon/xoff control
-	tty.c_cflag |= CLOCAL | CREAD;//ignore modem controls, enable reading
-	tty.c_cflag &= ~(PARENB | PARODD);//shut off parity
-	tty.c_cflag |= parity;//set parity
-	tty.c_cflag &= ~CSTOPB;
-	tty.c_cflag &= ~CRTSCTS;
-
-	if (tcsetattr(fd, TCSANOW, &tty) != 0) {
-		printf("error %d from tcsetattr\n", errno);
-		return -1;
-	}
-	return 0;
-}
-
-void set_blocking(int fd, bool should_block) {
-	struct termios tty;
-	memset(&tty, 0, sizeof(tty));
-	if (tcgetattr(fd, &tty) != 0) {
-		printf("error %d from tcgetattr\n", errno);
-	}
-
-	tty.c_cc[VMIN] = should_block ? 1 : 0;
-	tty.c_cc[VTIME] = 5;
-
-	if (tcsetattr(fd, TCSANOW, &tty) != 0) {
-		printf("error %d setting termios attrs\n", errno);
-	}
-}
-
-int open_com() {
-	string portname = "/dev/ttyUSB1";
-	int fd = open(portname.c_str(), O_RDWR | O_NOCTTY | O_SYNC);
-	if (fd < 0) {
-		printf("error %d opening %s: %s", errno, portname.c_str(), strerror(errno));
-	}
-
-	set_iface_attribs(fd, B115200, 0);
-	set_blocking(fd, 0);
-}
+int com_id;
 
 /**********
 public code
@@ -104,7 +45,7 @@ void initComms() {
 	e_pin.setPort(0);
 	e_pin.setDirection(INPUT);
 
-	curr_fd = open_com();
+	com_id = open_com("/dev/ttyUSB1");//TODO:fix
 }
 
 void setRead(int id) {
@@ -114,9 +55,9 @@ void setRead(int id) {
 	});
 }
 
-float read() {
-	char buf[100];
-	int n = read(curr_fd, buf, sizeof(buf));
+float readPort() {
+	int x = 0;
+	string output = read_port(com_id, 100);
 	//TODO: 4-byte float read protocol (check for terminating character)
 }
