+++ Pi/src/cpp/main.cpp
@@ -5,7 +5,11 @@
 #include "gpio/GPIOPin.h"
 
 int main() {
+#ifdef __arm__
 	initComms();
+#else
+	printf("Warning: probably not running on Pi, disabling GPIO and UART\n");
+#endif
 
 	return 0;
 }
